requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
constantly==15.1.0
cssselect==1.2.0
dateutils==0.6.12
EasyProcess==1.1
entrypoint2==1.1
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-headers==1.0.2
fake-useragent==1.2.1
filelock==3.12.2
Flask==3.0.0
frida==16.1.4
frida-tools==12.3.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
intensio-obfuscator==1.0.10.6
itchat==1.3.10
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keyboard==0.13.5
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
maxminddb==2.5.1
maxminddb-geolite2==2018.703
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parsel==1.8.1
pefile==2023.2.7
Pillow==10.0.0
playwright==1.38.0
progress==1.6
prompt-toolkit==3.0.39
Protego==0.3.0
pyaes==1.6.1
pyarmor.cli.core==5.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.19.0
PyDispatcher==2.0.7
pyee==9.0.4
PyExecJS2==1.6.1
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyperclip==1.8.2
pypng==0.20220715.0
PyQRCode==1.2.1
pyquery==2.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-obfuscator==0.0.2
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinauto==0.6.8
queuelib==1.6.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
retrying==1.3.4
rsa==4.9
Scrapy==2.10.0
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
Telethon==1.29.3
tldextract==3.4.4
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.2
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
